Update example, remove commented code
