+++ examples/with-wallet-stamper/src/components/account.tsx
@@ -25,7 +25,6 @@ export default function Account({
         <span className="text-center">{balance}</span>
       </Card>
       <Card className="w-min cursor-pointer rounded-lg p-2">
-        {/* <LogOut className="mr-2 h-4 w-4" /> */}
         {truncatedAddress}
       </Card>
       <Button onClick={() => disconnect()}>
